+++ config.ru
@@ -1,3 +1,8 @@
 require_relative 'time'
 
-run TimeApp.new
+ROUTES = {
+  '/time' => TimeApp.new
+}.freeze
+
+use Rack::ContentType, "text/plain"
+run Rack::URLMap.new(ROUTES)
